+++ .vscode-test.mjs
@@ -1,7 +1,7 @@
 import { defineConfig } from '@vscode/test-cli';
 
 export default defineConfig({
-	files: [
-		'out/test/**/*.test.js'
-	],
+  label: 'unitTests',
+  workspaceFolder: './src/test/TestWorkspace/',
+  files: 'out/test/**/*.test.js'
 });