+++ src/test/TestClasses/TestClass1.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace TestClasses;
-
-//This is a test class to test basic renaming.
-class TestClass1
-{
-}