+++ src/test/TestClasses/TestClass2.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace TestClasses;
-
-class TestClass2
-{
-}