+++ src/test/TestWorkspace/src/App/TestBase.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App;
+
+/**
+ * Used for checking renaming and moving
+ */
+class TestBase
+{
+  function testing(){
+    echo 'This is a test';
+  }
+}