+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to the "php-file-refactoring" extension will be documented i
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.1.4]
+
+- Fixed issue with updating own class when renaming.
+- Fixed use item alias issues that removed the alias.
+- Added custom composer.json location in configuration
+- Added option to automatically save files on changes
+- Added tests for functionality
+
 ## [0.1.3]
 
 - Added option to disable name type update in settings, if you have another plugin that updates the files type name (within it's self) this setting will stop this extension from updating it.